Extract common structs and functions into separate module
